+++ src/chapter2/ChapterTwo.java
@@ -1,4 +1,4 @@
-package chapterTwo;
+package chapter2;
 
 import java.util.HashSet;
 import java.util.LinkedList;
