+++ test/chapterThree/TestArrayThreeStack.java
@@ -4,6 +4,8 @@ import static org.junit.jupiter.api.Assertions.*;
 
 import org.junit.jupiter.api.Test;
 
+import chapter3.ArrayThreeStack;
+
 class TestArrayThreeStack {
 
 	@Test
