+++ test/chapterThree/TestStackMin.java
@@ -4,6 +4,8 @@ import static org.junit.jupiter.api.Assertions.*;
 
 import org.junit.jupiter.api.Test;
 
+import chapter3.StackMin;
+
 class TestStackMin {
 
 	@Test
